# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAP API",
    "version": "1.0.0",
    "description": "Kontrak rancangan SAP untuk frontend/backend. Kebutuhan otorisasi dan domain ada di API_SPEC.md. Belum merupakan layanan berjalan."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "Daftar dan kirim OTP",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nilai dari getCsrf; cocok cookie/token terverifikasi."
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted asynchronous work",
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "403": {
            "description": "Forbidden or CSRF invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "409": {
            "description": "Revision or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "413": {
            "description": "Image too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "415": {
            "description": "Unsupported image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "422": {
            "description": "Domain validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                },
                "description": "Detik sebelum retry."
              },
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          }
        },
        "x-required-role": "public",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInput"
              }
            }
          }
        }
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "verifyEmail",
        "summary": "Konsumsi OTP dan mulai sesi",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nilai dari getCsrf; cocok cookie/token terverifikasi."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              },
              "Set-Cookie": {
                "$ref": "#/components/headers/SetCookie"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "403": {
            "description": "Forbidden or CSRF invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "409": {
            "description": "Revision or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "413": {
            "description": "Image too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "415": {
            "description": "Unsupported image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "422": {
            "description": "Domain validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                },
                "description": "Detik sebelum retry."
              },
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          }
        },
        "x-required-role": "public",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyInput"
              }
            }
          }
        }
      }
    },
    "/auth/resend-verification": {
      "post": {
        "operationId": "resendVerification",
        "summary": "Kirim ulang OTP dengan cooldown",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nilai dari getCsrf; cocok cookie/token terverifikasi."
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted asynchronous work",
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "403": {
            "description": "Forbidden or CSRF invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "409": {
            "description": "Revision or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "413": {
            "description": "Image too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "415": {
            "description": "Unsupported image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "422": {
            "description": "Domain validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                },
                "description": "Detik sebelum retry."
              },
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          }
        },
        "x-required-role": "public",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailInput"
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Login akun terverifikasi",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nilai dari getCsrf; cocok cookie/token terverifikasi."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              },
              "Set-Cookie": {
                "$ref": "#/components/headers/SetCookie"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "403": {
            "description": "Forbidden or CSRF invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "409": {
            "description": "Revision or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "413": {
            "description": "Image too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "415": {
            "description": "Unsupported image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "422": {
            "description": "Domain validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                },
                "description": "Detik sebelum retry."
              },
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          }
        },
        "x-required-role": "public",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Meta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "requestId"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "error",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "sapaEnabled": {
            "type": "boolean",
            "description": "Preferensi teman virtual SAPA. Default true untuk akun baru; UI membaca ini saat bootstrap. Lihat addendum SAPA_ASSISTANT.md."
          }
        },
        "required": [
          "id",
          "displayName",
          "email",
          "role",
          "emailVerified",
          "sapaEnabled"
        ],
        "additionalProperties": false
      },
...
      "Challenge": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "retryAfterSeconds": {
            "type": "integer",
            "minimum": 0
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "challengeId",
          "expiresAt",
          "retryAfterSeconds",
          "message"
        ],
        "additionalProperties": false
      },
      "RegisterInput": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "displayName",
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "LoginInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "VerifyInput": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$"
          }
        },
        "required": [
          "challengeId",
          "code"
        ],
        "additionalProperties": false
      },
      "EmailInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ],
        "additionalProperties": false
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/User"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "ChallengeResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Challenge"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sap_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "deletionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sap_deletion"
      }
    },
    "headers": {
      "ContractVersion": {
        "description": "Contract version echoed on every response (success and error) so clients can detect contract drift.",
        "required": true,
        "schema": {
          "type": "string",
          "enum": [
            "1.0.0"
          ]
        }
      },
      "SetCookie": {
        "description": "Sets or clears a host-only cookie. All auth cookies are HttpOnly, Secure, SameSite=Lax, Path=/. sap_csrf carries the signed double-submit CSRF token (GET /auth/csrf; rotated on login/logout). sap_session is the opaque session token set on verify-email/login and cleared on logout, reset-password, and account deletion. sap_deletion is the account-deletion receipt set by DELETE /users/me. A single response may emit more than one Set-Cookie line.",
        "schema": {
          "type": "string"
        }
      }
    }
  }
}